backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "issue-tracker-backend"
authors = [
    {name = "Rolika"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all] (>=0.115.14,<0.116.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "email-validator (>=2.2.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "google-auth (>=2.40.3,<3.0.0)",
    "requests (>=2.32.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
